+++ src/config/variables.ts
@@ -33,3 +33,5 @@ export const geoindexService = import.meta.env.VITE_GEOINDEX
  * Should we invert logo colors in dark mode?
  */
 export const darkModeLogoStyle = import.meta.env.VITE_DARK_MODE_LOGO_STYLE
+
+export const defaultCommunityName: string = 'SolidCouch'