+++ src/pages/HospexSetup/types.ts
@@ -1,0 +1,9 @@
+export enum SetupStatusKey {
+  isMember = 'isMember',
+  isHospexProfile = 'isHospexProfile',
+  isPublicTypeIndex = 'isPublicTypeIndex',
+  isPrivateTypeIndex = 'isPrivateTypeIndex',
+  isInbox = 'isInbox',
+  isSimpleEmailNotifications = 'isSimpleEmailNotifications',
+  isEmailNotifications = 'isEmailNotifications',
+}