+++ src/pages/MyHosts.tsx
@@ -1,3 +1,0 @@
-export const MyHosts = () => {
-  return <>MyHosts</>
-}