+++ src/pages/Profile/Profile.tsx
@@ -3,12 +3,14 @@ import { ProtectedImg } from '@/components/ProtectedImg.tsx'
 import { useConfig } from '@/config/hooks'
 import { useProfile } from '@/hooks/data/useProfile'
 import { useAuth } from '@/hooks/useAuth'
+import { Trans, useLingui } from '@lingui/react/macro'
 import { FaPencilAlt } from 'react-icons/fa'
 import { useParams } from 'react-router-dom'
 import { ManageContact } from './ManageContact.tsx'
 import styles from './Profile.module.scss'
 
 export const Profile = () => {
+  const { t } = useLingui()
   const { communityId } = useConfig()
   const personId = useParams().id as string
   const auth = useAuth()
@@ -17,12 +19,14 @@ export const Profile = () => {
   const [profile] = useProfile(personId, communityId)
   const [myProfile] = useProfile(auth.webId as string, communityId)
 
+  const name = profile.name
+
   return (
     <div className={styles.container}>
       <ProtectedImg
         className={styles.photo}
         src={profile.photo}
-        alt={`Profile photo of ${profile.name}`}
+        alt={t`Profile photo of ${name}`}
         data-cy="profile-photo"
       />
       <header className={styles.name} data-cy="profile-name">
@@ -33,16 +37,16 @@ export const Profile = () => {
       </section>
       {isMe && (
         <ButtonLink secondary to="/profile/edit" data-cy="edit-profile-link">
-          <FaPencilAlt /> edit profile
+          <FaPencilAlt /> <Trans>edit profile</Trans>
         </ButtonLink>
       )}
       {isMe === false && (
         <ButtonLink secondary to={`/messages/${encodeURIComponent(personId)}`}>
-          Write a message
+          <Trans>Write a message</Trans>
         </ButtonLink>
       )}
       <ButtonLink tertiary to="contacts">
-        contacts
+        <Trans>contacts</Trans>
       </ButtonLink>
       {isMe === false && <ManageContact webId={personId} />}
       <section>
@@ -52,7 +56,9 @@ export const Profile = () => {
             highlighted={isMe ? [] : (myProfile.interests ?? [])}
           />
         ) : (
-          <Loading>loading interests...</Loading>
+          <Loading>
+            <Trans>loading interests...</Trans>
+          </Loading>
         )}
       </section>
     </div>